✨Version 2.2.0

+++ VTOLVR-ModLoader/MainWindow.xaml.cs
@@ -36,7 +36,7 @@ namespace VTOLVR_ModLoader
         private static string updatesFileTemp = @"\updates_TEMP.xml";
         private static string updatesURL = @"/files/updates.xml";
         private string url = @"https://vtolvr-mods.com";
-        private string root;
+        public static string root;
         private string vtolFolder;
 
         //Startup
@@ -60,6 +60,10 @@ namespace VTOLVR_ModLoader
         private int extractedMods = 0;
         private int extractedSkins = 0;
         private int movedDep = 0;
+        //Dev Console
+        public static bool devConsole;
+        //Settings
+        public static Settings settings;
 
         private static string CalculateMD5(string filename)
         {
@@ -307,8 +311,8 @@ namespace VTOLVR_ModLoader
             GifState(gifStates.Play);
 
             //Launching the game
-            /*
-            if (updateExe)
+            
+            if (devConsole)
             {
                 string regPath = (string)Registry.GetValue(
     @"HKEY_CURRENT_USER\SOFTWARE\Valve\Steam",
@@ -317,12 +321,12 @@ namespace VTOLVR_ModLoader
 
                 Process process = new Process();
                 process.StartInfo.FileName = regPath + @"\steam.exe";
-                process.StartInfo.Arguments = @"-applaunch 667970" + " -updateLauncher " + newExeVersion;
+                process.StartInfo.Arguments = @"-applaunch 667970" + " dev";
                 process.Start();
             }
                 
             else
-            */
+            
             Process.Start("steam://run/667970");
 
             //Searching For Process
@@ -500,13 +504,30 @@ namespace VTOLVR_ModLoader
                         {
                             split = depFiles[k].Split('\\');
                             fileName = split[split.Length - 1];
-                            Console.WriteLine("Moved file \n" + Directory.GetParent(root).FullName +
+
+                            if (File.Exists(Directory.GetParent(root).FullName +
+                                        @"\VTOLVR_Data\Managed\" + fileName))
+                            {
+                                string oldHash = CalculateMD5(Directory.GetParent(root).FullName +
                                         @"\VTOLVR_Data\Managed\" + fileName);
-                            File.Copy(depFiles[k], Directory.GetParent(root).FullName +
+                                string newHash = CalculateMD5(depFiles[k]);
+                                if (!oldHash.Equals(newHash))
+                                {
+                                    File.Copy(depFiles[k], Directory.GetParent(root).FullName +
                                         @"\VTOLVR_Data\Managed\" + fileName,
                                         true);
-
-                            movedDep++;
+                                    movedDep++;
+                                }
+                            }
+                            else
+                            {
+                                Console.WriteLine("Moved file \n" + Directory.GetParent(root).FullName +
+                                        @"\VTOLVR_Data\Managed\" + fileName);
+                                File.Copy(depFiles[k], Directory.GetParent(root).FullName +
+                                            @"\VTOLVR_Data\Managed\" + fileName,
+                                            true);
+                                movedDep++;
+                            }
                         }
                         break;
                     }
@@ -590,33 +611,17 @@ namespace VTOLVR_ModLoader
             Process.GetCurrentProcess().Kill();
         }
 
-        private void WebsiteMods(object sender, RoutedEventArgs e)
+        private void Website(object sender, RoutedEventArgs e)
         {
-            Process.Start("https://vtolvr-mods.com/mods.php");
+            Process.Start("https://vtolvr-mods.com");
         }
-
-        private void WebsiteSkins(object sender, RoutedEventArgs e)
-        {
-            Process.Start("https://vtolvr-mods.com/skins.php");
-        }
-
         private void Discord(object sender, RoutedEventArgs e)
         {
             Process.Start("https://discord.gg/49HDD7m");
         }
-
-        private void ModCreator(object sender, RoutedEventArgs e)
+        private void Patreon(object sender, RoutedEventArgs e)
         {
-            Mod newMod = new Mod();
-            newMod.name = "Mod Name";
-            newMod.description = "Mod Description";
-            using (FileStream stream = new FileStream(root + @"\info.xml", FileMode.Create))
-            {
-                XmlSerializer xml = new XmlSerializer(typeof(Mod));
-                xml.Serialize(stream, newMod);
-            }
-
-            MessageBox.Show("Created info.xml in \n\"" + root + "\"", "Created Info.xml", MessageBoxButton.OK, MessageBoxImage.Information);
+            Process.Start("https://www.patreon.com/vtolvrmods");
         }
 
         private void Quit(object sender, RoutedEventArgs e)
@@ -656,6 +661,19 @@ namespace VTOLVR_ModLoader
         }
 
         #endregion
+
+        private void OpenSettings(object sender, RoutedEventArgs e)
+        {
+            if (settings != null)
+            {
+                settings.Activate();
+                return;
+            }
+            settings = new Settings();
+            settings.Show();
+        }
+
+
     }
 
     public class Mod
@@ -663,5 +681,11 @@ namespace VTOLVR_ModLoader
         public string name;
         public string description;
         public Mod() { }
+
+        public Mod(string name, string description)
+        {
+            this.name = name;
+            this.description = description;
+        }
     }
 }
